updated the examples
